After flying around the world during October and November, I have a lot of new code to commit.
git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/pgtips-srv-and-outbound-stuff-1.8@402707 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ channels/sip/config_parser.c
@@ -641,7 +641,7 @@ alloc_fail:
 	return res;
 }
 
-int sip_parse_host(char *line, int lineno, char **hostname, int *portnum, enum sip_transport *transport)
+int sip_parse_host(char *line, int lineno, char **hostname, unsigned short *portnum, enum sip_transport *transport)
 {
 	char *port;
